refactor(hashtower): rename proof-builder.ts to hashtower-hash-chain-proof-builder.ts
