+++ packages/hashtower/src/hashtower-hash-chain-proof-builder.ts
