+++ packages/hashtower/src/index.ts
@@ -1,4 +1,4 @@
-import HashTowerHashChainProofBuilder from "./proof-builder"
+import HashTowerHashChainProofBuilder from "./hashtower-hash-chain-proof-builder"
 
 export { HashTowerHashChainProofBuilder }
 export * from "./types"
